+++ controllers/config/config.go
@@ -27,7 +27,7 @@ type ControllerConfig struct {
 	MaxRetainedPackagesPerApp        int                `yaml:"maxRetainedPackagesPerApp"`
 	MaxRetainedBuildsPerApp          int                `yaml:"maxRetainedBuildsPerApp"`
 	LogLevel                         zapcore.Level      `yaml:"logLevel"`
-	SpaceFinalizerAppDeletionTimeout *int64             `yaml:"spaceFinalizerAppDeletionTimeout"`
+	SpaceFinalizerAppDeletionTimeout *int32             `yaml:"spaceFinalizerAppDeletionTimeout"`
 
 	// job-task-runner
 	JobTTL                                     string `yaml:"jobTTL"`
@@ -51,7 +51,7 @@ type ControllerConfig struct {
 type CFProcessDefaults struct {
 	MemoryMB    int64  `yaml:"memoryMB"`
 	DiskQuotaMB int64  `yaml:"diskQuotaMB"`
-	Timeout     *int64 `yaml:"timeout"`
+	Timeout     *int32 `yaml:"timeout"`
 }
 
 type CFStagingResources struct {
@@ -67,7 +67,7 @@ type Networking struct {
 
 const (
 	defaultTaskTTL            = 30 * 24 * time.Hour
-	defaultTimeout      int64 = 60
+	defaultTimeout      int32 = 60
 	defaultJobTTL             = 24 * time.Hour
 	defaultBuildCacheMB       = 2048
 )
